+++ R/snpRdata.R
@@ -347,6 +347,8 @@ snpRdata <- setClass(Class = 'snpRdata', slots = c(sample.meta = "data.frame",
 #'  for each individual. For structure input files only.
 #'@param marker_names logical, default FALSE. If TRUE, assumes that a
 #'  header row of marker is present. For structure input files only.
+#'@param fix_overlaps Logical, default TRUE. If TRUE, overlapping positions will
+#'  be checked and fixed during 'ms' file import.
 #'@param verbose Logical, default FALSE. If TRUE, will print a few status
 #'  updates and checks.
 #'@param .pass_filters Internal, probably not for user use. Used to pass 
@@ -402,7 +404,8 @@ snpRdata <- setClass(Class = 'snpRdata', slots = c(sample.meta = "data.frame",
 #'
 #'@author William Hemstrom
 import.snpR.data <- function(genotypes, snp.meta = NULL, sample.meta = NULL, mDat = "NN", chr.length = NULL,
-                             ..., header_cols = 0, rows_per_individual = 2, marker_names = FALSE, verbose = FALSE,
+                             ..., header_cols = 0, rows_per_individual = 2, marker_names = FALSE, fix_overlaps = TRUE,
+                             verbose = FALSE,
                              .pass_filters = FALSE, .skip_filters = FALSE){
   position <- .snp.id <- .sample.id <- NULL
 
@@ -445,7 +448,11 @@ import.snpR.data <- function(genotypes, snp.meta = NULL, sample.meta = NULL, mDa
         return(.process_vcf(genotypes, snp.meta, sample.meta))
       }
       else if(grepl("\\.ms$", genotypes)){
-        return(format_snps(genotypes, input_format = "ms", snp.meta = snp.meta, sample.meta = sample.meta, chr.length = chr.length))
+        if(!is.null(snp.meta)){
+          snp.meta <- NULL
+          warning("Any provided snp.meta will be discarded during ms import.\n")
+        }
+        return(.process_ms(genotypes, chr.length, sample.meta, fix_overlaps))
       }
       else if(grepl("\\.genepop$", genotypes) | grepl("\\.gen$", genotypes)){
         return(.process_genepop(genotypes, snp.meta, sample.meta, mDat))
@@ -923,7 +930,7 @@ get.snpR.stats <- function(x, facets = NULL, stats = "single", bootstraps = FALS
       keep.cols <- colnames(y)[keep.cols]
       
       good.cols <- keep.cols[which(keep.cols %in% c("facet", "subfacet", "snp.facet", "snp.subfacet", "comparison", "pop", "sigma", "step", "nk.status", "gaussian", "triple_sigma", "n_snps",
-                                                    colnames(x@facet.meta)[-which(colnames(x@facet.meta) == ".snp.id")], colnames(sample.meta(x))))]
+                                                    colnames(x@facet.meta), colnames(sample.meta(x))))]
       grep.cols <- numeric(0)
       for(i in 1:length(col_pattern)){
         grep.cols <- c(grep.cols, grep(col_pattern[i], colnames(y)))
@@ -949,7 +956,7 @@ get.snpR.stats <- function(x, facets = NULL, stats = "single", bootstraps = FALS
       return(NULL)
     }
     
-    if(all(colnames(y)[keep.cols] %in% c("facet", "subfacet", "snp.facet", "snp.subfacet", "comparison"))){
+    if(all(colnames(y)[keep.cols] %in% c("facet", "subfacet", "snp.facet", "snp.subfacet", ".snp.id", "comparison"))){
       return(NULL)
     }
     return(as.data.frame(.fix..call(y[keep.rows, ..keep.cols]), stringsAsFactors = FALSE))